[measureobjectneighbors][functions]: PR review feedback implemented 1. Removed redundant imports 2. Awaiting reply to comment on replacing frequently used code block (for unpacking measurements to core) 3. replaced ternary if statement with `self.neighbors_name.value` as it achieves the same result 4. Removed utils functions to `get_nkept_objects()` and `get_kept_label_has_pixels()`, updated docs for ObjectSegmentations to clarify that object nos are contiguous and start at 1 5. Removed `def __str__(self):` from the enum class so that it is consistent with all other modules 6. Moved functions from `functions/object_processing.py` to `functions/segmentation.py` for consistency with the purpose of the function
