+++ src/subpackages/library/cellprofiler_library/functions/measurement.py
@@ -26,9 +26,8 @@ from cellprofiler_library.functions.segmentation import count_from_ijv
 from cellprofiler_library.functions.segmentation import areas_from_ijv
 from cellprofiler_library.functions.segmentation import cast_labels_to_label_set
 from cellprofiler_library.functions.segmentation import convert_label_set_to_ijv
-from cellprofiler_library.functions.segmentation import indices_from_ijv
 from cellprofiler_library.functions.image_processing import masked_erode, restore_scale, get_morphology_footprint
-from cellprofiler_library.functions.object_processing import relate_labels
+from cellprofiler_library.functions.segmentation import relate_labels
 
 from cellprofiler_library.types import Pixel, ObjectLabel, ImageGrayscale, ImageGrayscaleMask, ImageAny, ImageBinary, ImageBinaryMask, ObjectSegmentation, ObjectLabelsDense, ObjectLabelSet, ObjectSegmentationIJV, Image2DBinary, Image2DColor, Image2DGrayscale
 from cellprofiler_library.opts.objectsizeshapefeatures import ZERNIKE_N
