+++ src/subpackages/library/cellprofiler_library/opts/measureobjectneighbors.py
@@ -4,8 +4,6 @@ class DistanceMethod(str, Enum):
     ADJACENT = "Adjacent"
     EXPAND = "Expand until adjacent"
     WITHIN = "Within a specified distance"
-    def __str__(self):
-        return self.value
 
 class Measurement(str, Enum):
     NUMBER_OF_NEIGHBORS = "NumberOfNeighbors"
@@ -15,29 +13,25 @@ class Measurement(str, Enum):
     SECOND_CLOSEST_OBJECT_NUMBER = "SecondClosestObjectNumber"
     SECOND_CLOSEST_DISTANCE = "SecondClosestDistance"
     ANGLE_BETWEEN_NEIGHBORS = "AngleBetweenNeighbors"
-    def __str__(self):
-        return self.value
 
 D_ALL = [
-    DistanceMethod.ADJACENT,
-    DistanceMethod.EXPAND,
-    DistanceMethod.WITHIN,
+    DistanceMethod.ADJACENT.value,
+    DistanceMethod.EXPAND.value,
+    DistanceMethod.WITHIN.value,
 ]
 
 M_ALL = [
-    Measurement.NUMBER_OF_NEIGHBORS,
-    Measurement.PERCENT_TOUCHING,
-    Measurement.FIRST_CLOSEST_OBJECT_NUMBER,
-    Measurement.FIRST_CLOSEST_DISTANCE,
-    Measurement.SECOND_CLOSEST_OBJECT_NUMBER,
-    Measurement.SECOND_CLOSEST_DISTANCE,
-    Measurement.ANGLE_BETWEEN_NEIGHBORS,
+    Measurement.NUMBER_OF_NEIGHBORS.value,
+    Measurement.PERCENT_TOUCHING.value,
+    Measurement.FIRST_CLOSEST_OBJECT_NUMBER.value,
+    Measurement.FIRST_CLOSEST_DISTANCE.value,
+    Measurement.SECOND_CLOSEST_OBJECT_NUMBER.value,
+    Measurement.SECOND_CLOSEST_DISTANCE.value,
+    Measurement.ANGLE_BETWEEN_NEIGHBORS.value,
 ]
 
 class MeasurementScale(str, Enum):
     EXPANDED = "Expanded"
     ADJACENT = "Adjacent"
-    def __str__(self):
-        return self.value
 
 C_NEIGHBORS = "Neighbors"
